+++ tests/suites/0-create.ts
@@ -4,18 +4,18 @@ import fs from 'fs'
 import faker from 'faker'
 import nanoid from 'nanoid'
 
-import { Db, Table, primary, prop, Collection } from '../../src'
+import { Db, Table, primary, prop, Entity } from '../../src'
 
-@Table({ name: 'note', timestamp: true })
+@Entity({ name: 'note', timestamp: true })
 class DbNote {
   @prop({ unique: true }) key?: string
   @prop() data!: Record<string, any>
   @prop() order!: Record<string, number>
 }
 
-const dbNote = Collection.make(DbNote)
+const dbNote = new Table(DbNote)
 
-@Table({ name: 'card', timestamp: true })
+@Entity({ name: 'card', timestamp: true })
 class DbCard {
   @primary({ default: () => nanoid() }) _id?: string
   @prop({ references: dbNote }) noteId!: any
@@ -36,19 +36,19 @@ class DbCard {
   @prop({ type: 'StringArray', null: true }) v2b?: string[]
 }
 
-const dbCard = Collection.make(DbCard)
+const dbCard = new Table(DbCard)
 
-@Table({ name: 'media', timestamp: true })
+@Entity({ name: 'media', timestamp: true })
 class DbMedia {
   @prop() name!: string
   @prop() data!: ArrayBuffer
 }
 
-const dbMedia = Collection.make(DbMedia)
+const dbMedia = new Table(DbMedia)
 
 export async function initDatabase (filename: string = 'tests/test.db') {
   const db = await Db.connect(filename)
-  await Collection.init(db, [dbNote, dbCard, dbMedia])
+  await db.init([dbNote, dbCard, dbMedia])
 
   _db = {
     db,
@@ -71,7 +71,7 @@ export async function createDatabase (filename: string = 'tests/test.db') {
 
   await Promise.all(Array.from({ length: 1000 }).map(async () => {
     try {
-      await db.cols.media.create({
+      await db.db.create(dbMedia)({
         name: faker.system.fileName(undefined, 'image'),
         data: crypto.randomBytes(256),
       })
@@ -86,7 +86,7 @@ export async function createDatabase (filename: string = 'tests/test.db') {
         .map(() => [faker.hacker.noun(), faker.hacker.phrase()])
         .reduce((prev, [k, v]) => ({ ...prev, [k]: v }), {})
 
-      const noteId = await db.cols.note.create({
+      const noteId = await db.db.create(dbNote)({
         key: nanoid(),
         data,
         order: Object.keys(data).map((k, i) => [k, i]).reduce((prev, [k, i]) => ({ ...prev, [k]: i }), {}),
@@ -94,7 +94,7 @@ export async function createDatabase (filename: string = 'tests/test.db') {
 
       await Promise.all(Array.from({ length: faker.random.number(3) }).map(async () => {
         try {
-          await db.cols.card.create({
+          await db.db.create(dbCard)({
             isCool: faker.random.arrayElement([true, false]),
             v2b: faker.random.number(5) === 0
               ? undefined
@@ -134,9 +134,9 @@ export async function createDatabase (filename: string = 'tests/test.db') {
 let _db: {
   db: Db
   cols: {
-    note: Collection<DbNote>
-    media: Collection<DbMedia>
-    card: Collection<DbCard>
+    note: Table<DbNote>
+    media: Table<DbMedia>
+    card: Table<DbCard>
   }
 }
